Tighten SEO editor form submission handling
